Add symlink pointing to latest version.
When in doubt, we point here. Keep this up to date so that omv can have
a sensible default if none are manually specified. If we need to change
this semantic, we can do so at a later date. One possible change could
be to have a 'default' semantic, instead of 'latest'. This 'default'
semantic could be a configuration which is generally expected to work in
most instances. If we had a vagrant kubernetes provider, we could detect
which version and just act accordingly, automatically. That's a bigger
project for a later date.